+++ multitld/test/merge-score-test.cc
@@ -2,19 +2,19 @@
 
 int main(int argc, char** args) {
 
-    vector<float> scores1 = { 0.5f, 0.6f, 0.7f, 0.6f, 0.5f};
+    vector<double> scores1 = { 0.5, 0.6, 0.7, 0.6, 0.5};
     EnsembleScore* ensembleScore1 = new EnsembleScore(scores1);
 
-    vector<float> scores2 = { 0.3f, 0.2f, 0.1f, 0.2f, 0.75f};
+    vector<double> scores2 = { 0.3, 0.2, 0.1, 0.2, 0.75};
     EnsembleScore* ensembleScore2 = new EnsembleScore(scores2);
 
     EnsembleScore* ensembleScore3 = (EnsembleScore*) ensembleScore1->sum(ensembleScore2);
     ensembleScore3 = (EnsembleScore*) ensembleScore3->divide(2);
 
     for (int i = 0; i < 5; i++) {
-        float score1 = ensembleScore1->scores[i];
-        float score2 = ensembleScore2->scores[i];
-        float score3 = ensembleScore3->scores[i];
+        double score1 = ensembleScore1->scores[i];
+        double score2 = ensembleScore2->scores[i];
+        double score3 = ensembleScore3->scores[i];
 
         printf("%d, %f, %f, %f\n",i, score1, score2, score3);
     }
